+++ src/audit/message-delivery-progress-store.test.ts
@@ -9,6 +9,7 @@ import {
   closeOpenClawStateDatabaseForTest,
   openOpenClawStateDatabase,
 } from "../state/openclaw-state-db.js";
+import { STATE_SCHEMA_10_TO_9_DOWNGRADE_SQL } from "../state/openclaw-state-schema-v10-retirement.test-support.js";
 import { recordAuditEvent } from "./audit-event-store.js";
 import type { OutboundMessageProgressInput } from "./audit-event-types.js";
 import {
@@ -137,10 +138,10 @@ describe("outbound message progress companion", () => {
     );
   });
 
-  it("stays absent through startup, reads, and terminal-only writes at schema v9", () => {
+  it("stays absent through startup, reads, and terminal-only writes at schema v10", () => {
     const database = databaseOptions();
     const opened = openOpenClawStateDatabase(database);
-    expect(OPENCLAW_STATE_SCHEMA_VERSION).toBe(9);
+    expect(OPENCLAW_STATE_SCHEMA_VERSION).toBe(10);
     expect(tableExists(opened.db, "outbound_message_progress")).toBe(false);
     expect(tableExists(opened.db, "outbound_message_execution_bindings")).toBe(false);
 
@@ -274,6 +275,10 @@ describe("outbound message progress companion", () => {
     openOpenClawStateDatabase(database).db.exec(
       "ALTER TABLE skill_workshop_proposals DROP COLUMN claim_released_time;",
     );
+    // The pinned reader is a v9-era build, so it refuses v10 databases by the
+    // version contract. Project the file back to the exact v9 shape with the
+    // documented 10→9 downgrade so the C04 same-version reader proof survives.
+    openOpenClawStateDatabase(database).db.exec(STATE_SCHEMA_10_TO_9_DOWNGRADE_SQL);
     closeOpenClawStateDatabaseForTest();
 
     const repositoryRoot = process.cwd();
